+++ src/alpos/controller/BookController.java
@@ -91,3 +91,4 @@ public class BookController {
 		model.addAttribute("books", books);
 		return "books/index";
 	}
+}
