--- src/model/PublisherModel.java
@@ -1,19 +1,0 @@
-package model;
-
-public class PublisherModel {
-    private Integer id;
-    private String  name;
-    
-	public Integer getId() {
-		return id;
-	}
-	public void setId(Integer id) {
-		this.id = id;
-	}
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
-}
